Test _IS_ESCAPING_NEEDED properly
Since we use prysk to test itself, lines ending with "(esc)" need not
match verbatim. They'll also happily match lines that aren't escaped at
all as prysk.diff.esc first tries to unescape the expected line and
match it against the output, which thus need not be escaped.

The workaround is to use "(re)" to make sure there actually is escaping
and the "(esc)" annotation present in the output.